BackEnd/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
pydantic-settings
python-dotenv
